+++ src/templates/homepage/HeroSection/HeroButton.tsx
@@ -1,0 +1,23 @@
+import editorStyles from "styles/isomer-cms/pages/Editor.module.scss"
+
+import { getClassNames } from "templates/utils/stylingUtils"
+
+interface HeroButtonProps {
+  button?: string
+}
+export const HeroButton = ({ button }: HeroButtonProps) => (
+  <>
+    {button ? (
+      <div
+        className={getClassNames(editorStyles, [
+          "bp-button",
+          "is-uppercase",
+          "search-button",
+          "is-secondary",
+        ])}
+      >
+        {button}
+      </div>
+    ) : null}
+  </>
+)
